rearrange signing module #467
