+++ src/sign/mod.rs
@@ -13,7 +13,35 @@
 //! with Identifiable Aborts. [EPrint archive,
 //! 2021](https://eprint.iacr.org/2021/060.pdf).
 
-mod participant;
-mod share;
+mod interactive_sign;
+mod non_interactive_sign;
 
-pub use participant::Input;
+use k256::Scalar;
+pub use non_interactive_sign::participant::Input;
+use tracing::error;
+
+use crate::errors::{InternalError, Result};
+
+/// ECDSA signature on a message.
+///
+/// When generated by this library, the signature will be produced by the
+/// threshold ECDSA algorithm by Canetti et al.
+#[derive(Debug, PartialEq, Eq)]
+pub struct Signature(k256::ecdsa::Signature);
+
+impl Signature {
+    pub(super) fn try_from_scalars(r: Scalar, s: Scalar) -> Result<Self> {
+        Ok(Self(k256::ecdsa::Signature::from_scalars(r, s)
+            .map_err(|e| {
+                error!("Failed to generate `Signature` from `Scalar`s but they should be correctly formatted {e:?}");
+                InternalError::InternalInvariantFailed
+            })?
+        ))
+    }
+}
+
+impl AsRef<k256::ecdsa::Signature> for Signature {
+    fn as_ref(&self) -> &k256::ecdsa::Signature {
+        &self.0
+    }
+}
