+++ src/sign/non_interactive_sign/mod.rs
@@ -1,0 +1,2 @@
+pub(super) mod participant;
+mod share;
