+++ src/sign/non_interactive_sign/share.rs
@@ -1,6 +1,5 @@
 use k256::Scalar;
 use serde::{Deserialize, Serialize};
-use tracing::error;
 
 use crate::{
     errors::{InternalError, Result},
@@ -41,29 +40,5 @@ impl std::ops::Add<SignatureShare> for Scalar {
     type Output = Self;
     fn add(self, rhs: SignatureShare) -> Self::Output {
         self + rhs.0
-    }
-}
-
-/// ECDSA signature on a message.
-///
-/// When generated by this library, the signature will be produced by the
-/// threshold ECDSA algorithm by Canetti et al.
-#[derive(Debug, PartialEq, Eq)]
-pub struct Signature(k256::ecdsa::Signature);
-
-impl Signature {
-    pub(super) fn try_from_scalars(r: Scalar, s: Scalar) -> Result<Self> {
-        Ok(Self(k256::ecdsa::Signature::from_scalars(r, s)
-            .map_err(|e| {
-                error!("Failed to generate `Signature` from `Scalar`s but they should be correctly formatted {e:?}");
-                InternalError::InternalInvariantFailed
-            })?
-        ))
-    }
-}
-
-impl AsRef<k256::ecdsa::Signature> for Signature {
-    fn as_ref(&self) -> &k256::ecdsa::Signature {
-        &self.0
     }
 }
